more pythonic renaming
