+++ src/yadism/output.py
@@ -7,7 +7,7 @@ Output
     docs
 """
 
-from .structure_functions.EvaluatedStructureFunction import ESFResult
+from .structure_functions.esf_result import ESFResult
 
 
 class Output(dict):
@@ -19,7 +19,7 @@ class Output(dict):
     # TODO shift function somewhere else?
     # the other alternative is to shift this to an external module (and ouf of runner)
     # and import from there, here and in yadism.__init__
-    def apply_PDF(self, pdfs):
+    def apply_pdf(self, pdfs):
         # iterate
         ret: dict = {}
         for obs in self:
@@ -30,7 +30,7 @@ class Output(dict):
             ret[obs] = []
             for kin in self[obs]:
                 ret[obs].append(
-                    ESFResult.from_dict(kin).apply_PDF(self["xgrid"], self["xiF"], pdfs)
+                    ESFResult.from_dict(kin).apply_pdf(self["xgrid"], self["xiF"], pdfs)
                 )
         return ret
